requirements.txt
streamlit
wikipedia-api
openai